Fix linter errors, flatten components
