+++ saleor/static/dashboard-next/js/__tests__/components/FilterCard.test.tsx
@@ -1,10 +1,9 @@
 import * as React from "react";
-import { MemoryRouter } from "react-router-dom";
 import * as renderer from "react-test-renderer";
 
-import FilterCard from "../../components/cards/FilterCard";
+import FilterCard from "../../components/FilterCard";
 
-describe("<CategoryDetails />", () => {
+describe("<FilterCard />", () => {
   it("renders properly", () => {
     const component = renderer.create(<FilterCard handleClear={jest.fn()} />);
     expect(component).toMatchSnapshot();
