+++ saleor/static/dashboard-next/js/category/components/RootCategoryList.tsx
@@ -3,10 +3,10 @@ import Card from "material-ui/Card";
 import IconButton from "material-ui/IconButton";
 import * as React from "react";
 
+import CategoryList from "../../components/CategoryList";
 import PageHeader from "../../components/PageHeader";
+import { RootCategoryChildrenQuery } from "../../gql-types";
 import i18n from "../../i18n";
-import { RootCategoryChildrenQuery } from "../gql-types";
-import CategoryList from "./CategoryList";
 
 interface RootCategoryListProps {
   data: RootCategoryChildrenQuery;
