+++ saleor/static/dashboard-next/js/category/views/CategoryCreateForm.tsx
@@ -1,9 +1,7 @@
 import Card from "material-ui/Card";
-import Grid from "material-ui/Grid";
 import * as React from "react";
 import { Redirect } from "react-router-dom";
 
-import ErrorMessageCard from "../../components/cards/ErrorMessageCard";
 import { NavigatorLink } from "../../components/Navigator";
 import PageHeader from "../../components/PageHeader";
 import i18n from "../../i18n";
