+++ saleor/static/dashboard-next/js/components/PageHeader.tsx
@@ -1,11 +1,9 @@
 import ArrowBackIcon from "material-ui-icons/ArrowBack";
 import CloseIcon from "material-ui-icons/Close";
-import Divider from "material-ui/Divider";
 import IconButton from "material-ui/IconButton";
 import { withStyles } from "material-ui/styles";
 import Typography from "material-ui/Typography";
 import * as React from "react";
-import { Link } from "react-router-dom";
 import Skeleton from "./Skeleton";
 
 const decorate = withStyles(theme => ({
